Move background and popup modules into lib and components
