+++ entrypoints/popup.html
@@ -8,7 +8,7 @@
   <body>
     <div id="app"></div>
     <script type="module">
-      import Popup from "~/popup.svelte"
+      import Popup from "~/components/pages/Popup.svelte"
 
       if (typeof window !== "undefined") {
         const { mount } = await import("svelte")
