+++ lib/background.ts
